@@ -1,75 +1,83 @@
-// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-// Created by Sam Deane, 31/01/2018.
-// All code (c) 2018 - present day, Elegant Chaos Limited.
-// For licensing terms, see http://elegantchaos.com/license/liberal/.
-// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+// // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+// // Created by Sam Deane, 31/01/2018.
+// // All code (c) 2018 - present day, Elegant Chaos Limited.
+// // For licensing terms, see http://elegantchaos.com/license/liberal/.
+// // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
-#if !os(watchOS) && (!os(iOS) || targetEnvironment(simulator))
+// #if !os(watchOS) && (!os(iOS) || targetEnvironment(simulator))
 
-import Logger
-import XCTest
+//   import Logger
+//   import Testing
 
-public extension XCTestCase {
-    static let DefaultFatalErrorTimeout = 1.0
+//     static let DefaultFatalErrorTimeout = 1.0
 
-    /**
-      Assert that a fatal error has been reported via the Log Manager.
-     */
+//     /**
+//       Assert that a fatal error has been reported via the Log Manager.
+//      */
 
-    @discardableResult func XCTAssertFatalError(manager: Manager = Manager.shared, timeout: TimeInterval = XCTestCase.DefaultFatalErrorTimeout, testcase: @escaping () -> Void) -> Any? {
-        func unreachable() -> Never {
-            // run forever, to simulate a function that never returns
-            repeat {
-                RunLoop.current.run()
-            } while true
-        }
+//     @discardableResult public func XCTAssertFatalError(
+//       manager: Manager = Manager.shared,
+//       timeout: TimeInterval = XCTestCase.DefaultFatalErrorTimeout, testcase: @escaping () -> Void
+//     ) -> Any? {
+//       func unreachable() -> Never {
+//         // run forever, to simulate a function that never returns
+//         repeat {
+//           RunLoop.current.run()
+//         } while true
+//       }
 
-        let expectation = expectation(description: "expectingFatalError")
-        var fatalLogged: Any?
+//       let expectation = expectation(description: "expectingFatalError")
+//       var fatalLogged: Any?
 
-        let _ = manager.installFatalErrorHandler { logged, _, _, _ in
-            fatalLogged = logged
-            expectation.fulfill()
-            unreachable()
-        }
+//       let _ = manager.installFatalErrorHandler { logged, _, _, _ in
+//         fatalLogged = logged
+//         expectation.fulfill()
+//         unreachable()
+//       }
 
-        DispatchQueue.global(qos: .default).async(execute: testcase)
+//       DispatchQueue.global(qos: .default).async(execute: testcase)
 
-        wait(for: [expectation], timeout: timeout)
+//       wait(for: [expectation], timeout: timeout)
 
-        Manager.shared.resetFatalErrorHandler()
-        return fatalLogged
-    }
+//       Manager.shared.resetFatalErrorHandler()
+//       return fatalLogged
+//     }
 
-    /**
-     Assert that a fatal error has been reported via the Log Manager, and check
-     that the message/object logged matches an expected value.
-     */
+//     /**
+//      Assert that a fatal error has been reported via the Log Manager, and check
+//      that the message/object logged matches an expected value.
+//      */
 
-    func XCTAssertFatalError<T: Equatable>(equals: T, timeout: TimeInterval = XCTestCase.DefaultFatalErrorTimeout, testcase: @escaping () -> Void) {
-        let result = XCTAssertFatalError(timeout: timeout, testcase: testcase)
-        guard let error = result as? T else {
-            XCTFail("unexpected message type: \(String(describing: result))")
-            return
-        }
+//     public func XCTAssertFatalError<T: Equatable>(
+//       equals: T, timeout: TimeInterval = XCTestCase.DefaultFatalErrorTimeout,
+//       testcase: @escaping () -> Void
+//     ) {
+//       let result = XCTAssertFatalError(timeout: timeout, testcase: testcase)
+//       guard let error = result as? T else {
+//         XCTFail("unexpected message type: \(String(describing: result))")
+//         return
+//       }
 
-        XCTAssertEqual(error, equals)
-    }
+//       XCTAssertEqual(error, equals)
+//     }
 
-    /**
-     Assert that a fatal error has been reported via the Log Manager, and check
-     that a test passes.
-     */
+//     /**
+//      Assert that a fatal error has been reported via the Log Manager, and check
+//      that a test passes.
+//      */
 
-    func XCTAssertFatalError<T>(testing: (T) -> Bool, timeout: TimeInterval = XCTestCase.DefaultFatalErrorTimeout, testcase: @escaping () -> Void) {
-        let result = XCTAssertFatalError(timeout: timeout, testcase: testcase)
-        guard let error = result as? T else {
-            XCTFail("unexpected message type: \(String(describing: result))")
-            return
-        }
+//     public func XCTAssertFatalError<T>(
+//       testing: (T) -> Bool, timeout: TimeInterval = XCTestCase.DefaultFatalErrorTimeout,
+//       testcase: @escaping () -> Void
+//     ) {
+//       let result = XCTAssertFatalError(timeout: timeout, testcase: testcase)
+//       guard let error = result as? T else {
+//         XCTFail("unexpected message type: \(String(describing: result))")
+//         return
+//       }
 
-        XCTAssertTrue(testing(error))
-    }
-}
+//       XCTAssertTrue(testing(error))
+//     }
+//   }
 
-#endif
+// #endif
