@@ -4,7 +4,9 @@
 // For licensing terms, see http://elegantchaos.com/license/liberal/.
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
+import Combine
 import Foundation
+import Synchronization
 
 /**
  Represents a channel through which log messages can be sent.
@@ -12,156 +14,161 @@ import Foundation
  to one or more handlers.
  */
 
-public class Channel {
-    /**
-     Default log handler which prints to standard out,
-     without appending the channel details.
-     */
-
-    public static let stdoutHandler = PrintHandler("default", showName: false, showSubsystem: false)
-
-    /**
-     Default handler to use for channels if nothing else is specified.
-
-     On the Mac this is an OSLogHandler, which will log directly to the console without
-     sending output to stdout/stderr.
-
-     On Linux it is a PrintHandler which will log to stdout.
-     */
-
-    static func initDefaultHandler() -> Handler {
-        #if os(macOS) || os(iOS)
-        if #available(macOS 10.12, iOS 10.0, *) {
-            return OSLogHandler("default")
-        }
-        #endif
-
-        return stdoutHandler // TODO: should perhaps be stderr instead?
+public actor Channel {
+  /// Name of the channel.
+  nonisolated public let name: String
+  
+  /// Subsytem of the channel.
+  nonisolated public let subsystem: String
+  
+  /// Is this channel enabled?
+  nonisolated public var enabled: Bool {
+    get { _enabled.load(ordering: .relaxed) }
+    set {
+      _enabled.store(newValue, ordering: .relaxed)
+      Task { await manager.channelUpdated(self) }
     }
+  }
 
-    public static let defaultHandler = initDefaultHandler()
+  /// Underlying atomic storage for the enabled flag.
+  private let _enabled: Atomic<Bool>
 
-    /**
-     Default subsystem if nothing else is specified.
-     If the channel name is in dot syntax (x.y.z), then the last component is
-     assumed to be the name, and the rest is assumed to be the subsystem.
-
-     If there are no dots, it's all assumed to be the name, and this default
-     is used for the subsytem.
-     */
-
-    static let defaultSubsystem = "com.elegantchaos.logger"
-
-    /**
-     Default log channel that clients can use to log their actual output.
-     This is intended as a convenience for command line programs which actually want to produce output.
-     They could of course just use print() for this (producing normal output isn't strictly speaking
-     the same as logging), but this channel exists to allow output and logging to be treated in a uniform
-     way.
-
-     Unlike most channels, we want this one to default to always being on.
-     */
-
-    public static let stdout = Channel("stdout", handlers: [stdoutHandler], alwaysEnabled: true)
-
-    public let name: String
-    public let subsystem: String
-    public var enabled: Bool
-
-    public var fullName: String {
-        "\(subsystem).\(name)"
+  /// Manager that this channel is registered with.
+  let manager: Manager
+  
+  /// Handler attached to this channel.
+  let handler: Handler
+  
+  /// Sequence of logged items.
+  let sequence = YieldingSequence<LoggedItem>()
+  
+  ///   Default subsystem if nothing else is specified.
+  ///   If the channel name is in dot syntax (x.y.z), then the last component is
+  ///   assumed to be the name, and the rest is assumed to be the subsystem.///
+  ///   If there are no dots, it's all assumed to be the name, and this default
+  ///   is used for the subsytem.
+  static let defaultSubsystem = "com.elegantchaos.logger"
+  
+  /// Initialise the channel.
+  public init(
+    _ name: String, handler: Handler? = nil,
+    alwaysEnabled: Bool = false, manager: Manager? = nil, autoRun: Bool = true
+  ) {
+    let manager = manager ?? Manager.shared
+    let components = name.split(separator: ".")
+    let last = components.count - 1
+    let shortName: String
+    let subName: String
+    
+    if last > 0 {
+      shortName = String(components[last])
+      subName = components[..<last].joined(separator: ".")
+    } else {
+      shortName = name
+      subName = Channel.defaultSubsystem
     }
-
-    let manager: Manager
-    var handlers: [Handler] = []
-
-    public init(_ name: String, handlers: @autoclosure () -> [Handler] = [defaultHandler], alwaysEnabled: Bool = false, manager: Manager = Manager.shared) {
-        let components = name.split(separator: ".")
-        let last = components.count - 1
-        let shortName: String
-        let subName: String
-
-        if last > 0 {
-            shortName = String(components[last])
-            subName = components[..<last].joined(separator: ".")
-        } else {
-            shortName = name
-            subName = Channel.defaultSubsystem
-        }
-
-        let fullName = "\(subName).\(shortName)"
-        let enabledList = manager.channelsEnabledInSettings
-
-        self.name = shortName
-        self.subsystem = subName
-        self.manager = manager
-        self.enabled = enabledList.contains(shortName) || enabledList.contains(fullName) || alwaysEnabled
-        self.handlers = handlers() // TODO: does this need to be a closure any more?
-
-        manager.register(channel: self)
+    
+    let fullName = "\(subName).\(shortName)"
+    let enabledList = manager.channelsEnabledInSettings
+    let isEnabled = enabledList.contains(shortName) || enabledList.contains(fullName) || alwaysEnabled
+    
+    self.name = shortName
+    self.subsystem = subName
+    self.manager = manager
+    self._enabled = .init(isEnabled)
+    self.handler = handler ?? Manager.defaultHandler
+    
+    Task {
+      await manager.run(channel: self)
     }
-
-    /**
-     Log something.
-
-     The logged value is an autoclosure, to avoid doing unnecessary work if the channel is disabled.
-
-     If the channel is enabled, we capture the logged value in the calling thread, by evaluating the autoclosure.
-     We then log the value asynchronously. The log manager serialises the logging, to avoid race conditions.
-
-     Note that reading the `enabled` flag is not serialised, to avoid taking unnecessary locks. It's theoretically
-     possible for another thread to be writing to this flag whilst we're reading it, if the channel state is being
-     changed. Thread sanitizer might flag this up, but it's harmless, and should generally only happen in a debug
-     setting.
-     */
-
-    public func log(_ logged: @autoclosure () -> Any, file: StaticString = #file, line: UInt = #line, column: UInt = #column, function: StaticString = #function) {
-        if enabled {
-            let value = logged()
-            manager.queue.async {
-                let context = Context(file: file, line: line, column: column, function: function)
-                self.handlers.forEach { $0.log(channel: self, context: context, logged: value) }
-            }
-        }
+  }
+  
+  /**
+   Log something.
+   
+   The logged value is an autoclosure, to avoid doing unnecessary work if the channel is disabled.
+   
+   If the channel is enabled, we capture the logged value, by evaluating the autoclosure.
+   We then log the value asynchronously.
+   */
+  
+  nonisolated public func log<T>(
+    _ logged: @autoclosure () -> T, file: StaticString = #file, line: UInt = #line,
+    column: UInt = #column, function: StaticString = #function, dso: UnsafeRawPointer = #dsohandle
+  ) {
+    if enabled {
+      let context = Context(
+        channel: self,
+        file: file, line: line, column: column, function: function, dso: dso)
+      let value = asSendable(logged)
+      sequence.yield(LoggedItem(value: value, context: context))
     }
-
-    public func debug(_ logged: @autoclosure () -> Any, file: StaticString = #file, line: UInt = #line, column: UInt = #column, function: StaticString = #function) {
-        #if DEBUG
-        if enabled {
-            log(logged(), file: file, line: line, column: column, function: function)
-        }
-        #endif
+  }
+  
+  nonisolated public func debug<T>(
+    _ logged: @autoclosure () -> T, file: StaticString = #file, line: UInt = #line,
+    column: UInt = #column, function: StaticString = #function, dso: UnsafeRawPointer = #dsohandle
+  ) {
+#if DEBUG
+    if enabled {
+      let context = Context(
+        channel: self,
+        file: file, line: line, column: column, function: function, dso: dso)
+      let value = asSendable(logged)
+      sequence.yield(LoggedItem(value: value, context: context))
     }
-
-    public func fatal(_ logged: @autoclosure () -> Any, file: StaticString = #file, line: UInt = #line, column: UInt = #column, function: StaticString = #function) -> Never {
-        let value = logged()
-        log(value, file: file, line: line, column: column, function: function)
-        manager.fatalHandler(value, self, file, line)
+#endif
+  }
+  
+  nonisolated public func fatal(
+    _ logged: @autoclosure () -> Any, file: StaticString = #file, line: UInt = #line,
+    column: UInt = #column, function: StaticString = #function
+  ) -> Never {
+    let value = logged()
+    log(value, file: file, line: line, column: column, function: function)
+    manager.fatalHandler(value, self, file, line)
+  }
+  
+  /// Kick off the handler. It will pull items from the
+  /// stream until it is finished.
+  /// NB: Under normal conditions this function is called automatically
+  /// when the channel is created. For testing purposes it's useful to
+  /// be able to call it manually, in order to wait for all logged items
+  /// to be processed.
+  internal func run() async {
+    for await item in sequence {
+      await handler.log(item)
     }
-}
-
-extension Channel: Hashable {
-    // For now, we treat channels with the same name as equal,
-    // as long as they belong to the same manager.
-    public func hash(into hasher: inout Hasher) {
-        name.hash(into: &hasher)
-    }
-
-    public static func == (lhs: Channel, rhs: Channel) -> Bool {
-        (lhs.name == rhs.name) && (lhs.manager === rhs.manager)
-    }
+  }
+  
+  /// Shut down the channel.
+  /// Logging items to the channel after this call
+  /// will do nothing.
+  /// NB: Under normal conditions this function does not need to be called.
+  public func shutdown() {
+    print("shutdown \(name)")
+    sequence.continuation.finish()
+  }
 }
 
 extension Channel: Identifiable {
-    public var id: String { fullName }
+  nonisolated public var id: String { "\(subsystem).\(name)" }
 }
 
-// MARK: Deprecated API
+extension Channel: Hashable {
+  // For now, we treat channels with the same name as equal,
+  // as long as they belong to the same manager.
+  nonisolated public func hash(into hasher: inout Hasher) {
+    subsystem.hash(into: &hasher)
+    name.hash(into: &hasher)
+  }
 
-public extension Channel {
-    @available(*, deprecated, message: "Use Manager.shared instead")
-    static var defaultManager: Manager { Manager.shared }
+  public static func == (lhs: Channel, rhs: Channel) -> Bool {
+    (lhs.subsystem == rhs.subsystem) &&
+    (lhs.name == rhs.name) &&
+    (lhs.manager === rhs.manager)
+  }
 }
 
-@available(*, deprecated, message: "Use ``Channel`` instead.")
-public typealias Logger = Channel
+func asSendable<T>(_ value: () -> T) -> Sendable where T: Sendable { value() }
+func asSendable<T>(_ value: () -> T) -> Sendable { String(describing: value()) }
