@@ -4,26 +4,29 @@
 // For licensing terms, see http://elegantchaos.com/license/liberal/.
 // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
-/**
- Encapsulates the context (function, line number, etc) in which the log statement was made.
- */
+/// Encapsulates the context (function, line number, etc) in which the log statement was made.
+public struct Context: CustomStringConvertible, Sendable {
+  let channel: Channel
+  let file: StaticString
+  let line: UInt
+  let function: StaticString
+  let column: UInt
+  let dso: UInt
 
-public struct Context: CustomStringConvertible {
-    let file: StaticString
-    let line: UInt
-    let function: StaticString
-    let column: UInt
-    let dso: UnsafeRawPointer
+  public init(
+    channel: Channel,
+    file: StaticString = #file, line: UInt = #line, column: UInt = #column,
+    function: StaticString = #function, dso: UnsafeRawPointer = #dsohandle
+  ) {
+    self.channel = channel
+    self.file = file
+    self.line = line
+    self.function = function
+    self.column = column
+    self.dso = UInt(bitPattern: dso)
+  }
 
-    public init(file: StaticString = #file, line: UInt = #line, column: UInt = #column, function: StaticString = #function, dso: UnsafeRawPointer = #dsohandle) {
-        self.file = file
-        self.line = line
-        self.function = function
-        self.column = column
-        self.dso = dso
-    }
-
-    public var description: String {
-        "\(file): \(line),\(column) - \(function)"
-    }
+  public var description: String {
+    "\(file): \(line),\(column) - \(function)"
+  }
 }
